+++ src/Symfony/Component/Mime/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
  * Support `binary` as a `Content-Transfer-Encoding`
  * Add PGP/MIME signing and encryption support with the `PgpSigner` and `PgpEncrypter` classes
  * Add `AbstractPart::setContentTypeParameter()`
+ * Add the `Group` class to put a group of mailboxes in a mailbox list header (RFC 5322 and RFC 6854)
+ * Add `MailboxListHeader::getAddressList()` and `MailboxListHeader::createAddressList()`
 
 8.0
 ---